+++ NEWS.md
@@ -1,6 +1,8 @@
 # version 0.9-8
 
-* allow to `st_cast` COMPOUNDCURVE, MULTISURFACE or CURVEPOLYGON to GEOMETRYCOLLECTION (and back); 
+* `st_inscribed_circle' computes the maximum inscribed circle for polygons (requires GEOS >= 3.9.0)
+
+* allow to `st_cast` COMPOUNDCURVE, MULTISURFACE or CURVEPOLYGON to GEOMETRYCOLLECTION (and back); #1573
 
 # version 0.9-7